ref: move models to pkg
This makes it possible for other projects to use the same model
definitions when making requests to the API.
