+++ internal/handler/endpoint_test.go
@@ -5,8 +5,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/crazybolillo/eryth/internal/model"
 	"github.com/crazybolillo/eryth/internal/service"
+	"github.com/crazybolillo/eryth/pkg/model"
 	"github.com/jackc/pgx/v5"
 	"net/http"
 	"net/http/httptest"
