+++ internal/service/bouncer.go
@@ -3,8 +3,8 @@ package service
 import (
 	"context"
 	"github.com/crazybolillo/eryth/internal/db"
-	"github.com/crazybolillo/eryth/internal/model"
 	"github.com/crazybolillo/eryth/internal/sqlc"
+	"github.com/crazybolillo/eryth/pkg/model"
 	"log/slog"
 )
